Several moves/renames
- `Handler` code to `package.scala`.
- `handle` -> `run`.
- `Event` trait to top level.
- Capitalize filenames.
